Return runtime errors from generated Start function
Doing this means that errors generated by user-written functions that cligen is generating a CLI for will be passed through to the user of the package.

This will allow the user to handle those errors in any manner they see fit, as opposed to forcing them into using the error handling that will be bundled with cligen.